Graphics,Rendering and render graphs overhaul

+++ src/Runtime/Resource/Mesh/MeshResource.cpp
@@ -2,6 +2,8 @@
 #include <Runtime/Graphics/GraphicsDeviceAPI.h>
 #include <Runtime/Resource/Mesh/MeshUtils.h>
 #include <Runtime/Graphics/Buffer/GraphicsBufferUpdateDesc.h>
+#include <Runtime/Graphics/Command/CommandList.h>
+
 namespace Portakal
 {
 	MeshResource::MeshResource(GraphicsDevice* pOwnerDevice)
@@ -13,6 +15,8 @@ namespace Portakal
 		mVertexSubItemSize = 0;
 		mIndexSubItemCount = 0;
 		mIndexType = MeshIndexType::Unknown;
+
+		mCmdList = mOwnerDevice->CreateCommandList({});
 	}
 	MeshResource::MeshResource()
 	{
@@ -23,17 +27,66 @@ namespace Portakal
 		mVertexSubItemSize = 0;
 		mIndexSubItemCount = 0;
 		mIndexType = MeshIndexType::Unknown;
+
+		mCmdList = mOwnerDevice->CreateCommandList({});
 	}
 	MeshResource::~MeshResource()
 	{
 		mOwnerDevice = nullptr;
 	}
-	void MeshResource::AllocateVertexes(const Byte* pData, const unsigned long long subItemSize, const unsigned long long subItemCount, const Array<InputElementDesc>& inputElements)
+	unsigned long long MeshResource::GetVertexBufferItemCount() const noexcept
 	{
-		DeleteVertexBuffer();
+		if (mVertexBuffer == nullptr)
+			return 0;
+
+		return mVertexBuffer->GetSubItemCount();
+	}
+	unsigned long long MeshResource::GetVertexBufferItemSize() const noexcept
+	{
+		if (mVertexBuffer == nullptr)
+			return 0;
+
+		return mVertexBuffer->GetSubItemSize();
+	}
+	unsigned long long MeshResource::GetIndexBufferItemCount() const noexcept
+	{
+		if (mIndexBuffer == nullptr)
+			return 0;
+
+		return mIndexBuffer->GetSubItemCount();
+	}
+	unsigned long long MeshResource::GetIndexBufferItemSize() const noexcept
+	{
+		if (mIndexBuffer == nullptr)
+			return 0;
+
+		return mIndexBuffer->GetSubItemSize();
+	}
+	void MeshResource::AllocateVertexes(const unsigned long long subItemSize, const unsigned long long subItemCount, const InputLayoutDesc& inputLayout)
+	{
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot allocate vertexes due to this resource is destroyed");
+			return;
+		}
+
+		/*
+		* Validate owner device
+		*/
 		if (mOwnerDevice == nullptr)
 			return;
 
+		/*
+		* Delete the former vertex buffer
+		*/
+		DeleteVertexBuffer();
+
+		/*
+		* Create vertex buffer
+		*/
 		GraphicsBufferCreateDesc bufferDesc = {};
 		bufferDesc.Type = GraphicsBufferType::VertexBuffer;
 		bufferDesc.SubItemCount = subItemCount;
@@ -43,14 +96,71 @@ namespace Portakal
 
 		mVertexSubItemCount = subItemCount;
 		mVertexSubItemSize = subItemSize;
-		mInputElements = inputElements;
+		mInputLayout = inputLayout;
 	}
-	void MeshResource::AllocateIndexes(const Byte* pData, const MeshIndexType indexType, const unsigned long long subItemCount)
+	void MeshResource::AllocateVertexes(const unsigned long long subItemSize, const unsigned long long subItemCount)
 	{
-		DeleteIndexBuffer();
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot allocate vertexes due to this resource is destroyed");
+			return;
+		}
+
+		/*
+		* Validate owner device
+		*/
 		if (mOwnerDevice == nullptr)
 			return;
 
+		if (mInputLayout.Elements.GetCursor() == 0)
+		{
+			LOG("MeshResource", "Cannot allocate vertexes due to there is no defined input elements yet. You should first use the method override of AllocateVertexes which accepts input element desc array");
+			return;
+		}
+
+		/*
+		* Delete the former vertex buffer
+		*/
+		DeleteVertexBuffer();
+
+		/*
+		* Create vertex buffer
+		*/
+		GraphicsBufferCreateDesc bufferDesc = {};
+		bufferDesc.Type = GraphicsBufferType::VertexBuffer;
+		bufferDesc.SubItemCount = subItemCount;
+		bufferDesc.SubItemSize = subItemSize;
+
+		mVertexBuffer = mOwnerDevice->CreateBuffer(bufferDesc);
+
+		mVertexSubItemCount = subItemCount;
+		mVertexSubItemSize = subItemSize;
+	}
+	void MeshResource::AllocateIndexes(const MeshIndexType indexType, const unsigned long long subItemCount)
+	{
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot allocate indexes due to this resource is destroyed");
+			return;
+		}
+
+		if (mOwnerDevice == nullptr)
+			return;
+
+		/*
+		* Delete the former index buffer
+		*/
+		DeleteIndexBuffer();
+
+		/*
+		* Create index buffer
+		*/
 		GraphicsBufferCreateDesc bufferDesc = {};
 		bufferDesc.Type = GraphicsBufferType::IndexBuffer;
 		bufferDesc.SubItemCount = subItemCount;
@@ -63,35 +173,161 @@ namespace Portakal
 	}
 	void MeshResource::UpdateVertexes(const Byte* pData, const unsigned long long offset, const unsigned long long size)
 	{
-		if (mOwnerDevice == nullptr || mVertexBuffer == nullptr)
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource is destroyed");
 			return;
+		}
 
+		/*
+		* Validate device and vertex buffer
+		*/
+		if (mOwnerDevice == nullptr)
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource does not have a valid owner GraphicsDevice");
+			return;
+		}
+		if (mVertexBuffer == nullptr || mVertexBuffer->IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource does not have a valid vertex buffer created with AllocateVertexes before");
+			return;
+		}
+	
+		/*
+		* Update the buffer
+		*/
 		GraphicsBufferUpdateDesc desc = {};
 		desc.pData = pData;
 		desc.Offset = offset;
 		desc.Size = size;
 
-		mOwnerDevice->UpdateBuffer(mVertexBuffer, desc);
+		mCmdList->Lock();
+		mCmdList->UpdateBuffer(desc, mVertexBuffer);
+		mCmdList->Unlock();
 	}
 	void MeshResource::UpdateIndexes(const Byte* pData, const unsigned long long offset, const unsigned long long size)
 	{
-		if (mOwnerDevice == nullptr || mIndexBuffer == nullptr)
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource is destroyed");
 			return;
+		}
+
+		/*
+		* Validate device and vertex buffer
+		*/
+		if (mOwnerDevice == nullptr)
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource does not have a valid owner GraphicsDevice");
+			return;
+		}
+		if (mVertexBuffer == nullptr || mVertexBuffer->IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource does not have a valid vertex buffer created with AllocateIndexes before");
+			return;
+		}
 
 		GraphicsBufferUpdateDesc desc = {};
 		desc.pData = pData;
 		desc.Offset = offset;
 		desc.Size = size;
 
-		mOwnerDevice->UpdateBuffer(mIndexBuffer, desc);
+		/*
+		* Update the buffer
+		*/
+		mCmdList->Lock();
+		mCmdList->UpdateBuffer(desc, mIndexBuffer);
+		mCmdList->Unlock();
+	}
+	void MeshResource::UpdateVertexes(const Byte* pData, const unsigned long long offset, const unsigned long long size, CommandList* pTargetCmdList)
+	{
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource is destroyed");
+			return;
+		}
 
+		/*
+		* Validate device and vertex buffer
+		*/
+		if (mOwnerDevice == nullptr)
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource does not have a valid owner GraphicsDevice");
+			return;
+		}
+		if (mVertexBuffer == nullptr || mVertexBuffer->IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update vertexes due to this resource does not have a valid vertex buffer created with AllocateVertexes before");
+			return;
+		}
+
+		/*
+		* Update the buffer
+		*/
+		GraphicsBufferUpdateDesc desc = {};
+		desc.pData = pData;
+		desc.Offset = offset;
+		desc.Size = size;
+
+		pTargetCmdList->UpdateBuffer(desc, mVertexBuffer);
+	}
+	void MeshResource::UpdateIndexes(const Byte* pData, const unsigned long long offset, const unsigned long long size, CommandList* pTargetCmdList)
+	{
+		/*
+		* Validate if destroyed
+		*/
+		if (IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource is destroyed");
+			return;
+		}
+
+		/*
+		* Validate device and vertex buffer
+		*/
+		if (mOwnerDevice == nullptr)
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource does not have a valid owner GraphicsDevice");
+			return;
+		}
+		if (mVertexBuffer == nullptr || mVertexBuffer->IsDestroyed())
+		{
+			LOG("MeshResource", "Cannot update indexes due to this resource does not have a valid vertex buffer created with AllocateIndexes before");
+			return;
+		}
+
+		/*
+		* Update the buffer
+		*/
+		GraphicsBufferUpdateDesc desc = {};
+		desc.pData = pData;
+		desc.Offset = offset;
+		desc.Size = size;
+
+		pTargetCmdList->UpdateBuffer(desc, mIndexBuffer);
+	}
+	void MeshResource::RegisterStateChangedDelegate(const Delegate<void, MeshResource*>& del)
+	{
+		mStateChangedEvent += del;
+	}
+	void MeshResource::RemoveStateChangedDelegate(const Delegate<void, MeshResource*>& del)
+	{
+		mStateChangedEvent -= del;
 	}
 	void MeshResource::DeleteVertexBuffer()
 	{
 		if (mVertexBuffer == nullptr || mOwnerDevice == nullptr)
 			return;
 
-		mInputElements.Clear();
 		mVertexBuffer->Destroy();
 		mVertexBuffer = nullptr;
 
@@ -109,11 +345,38 @@ namespace Portakal
 		mIndexSubItemCount = 0;
 		mIndexType = MeshIndexType::Unknown;
 	}
+	void MeshResource::SignalStateChangedEvent()
+	{
+		mStateChangedEvent.Invoke(this);
+	}
 	void MeshResource::DestroyCore()
 	{
+		/*
+		* Signal all the listeners
+		*/
+		SignalStateChangedEvent();
+
+		/*
+		* Clear the state changed event invocation list
+		*/
+		mStateChangedEvent.ClearInvocationList();
+
+		/*
+		* Delete the buffers
+		*/
 		DeleteVertexBuffer();
 		DeleteIndexBuffer();
 
+		/*
+		* Delete cmd list
+		*/
+		if(mCmdList == nullptr)
+			mCmdList->Destroy();
+		mCmdList = nullptr;
+
+		/*
+		* Reset the member variables
+		*/
 		mOwnerDevice = nullptr;
 	}
 }
